+++ tests/test_actions.py
@@ -11,7 +11,7 @@ from rasa_core.actions.action import (
     ActionExecutionRejection, ACTION_LISTEN_NAME, ACTION_RESTART_NAME,
     ACTION_DEFAULT_FALLBACK_NAME, ACTION_DEACTIVATE_FORM_NAME,
     ACTION_REVERT_FALLBACK_EVENTS_NAME, ACTION_DEFAULT_ASK_AFFIRMATION_NAME,
-    ACTION_DEFAULT_ASK_REPHRASE_NAME)
+    ACTION_DEFAULT_ASK_REPHRASE_NAME, ACTION_BACK_NAME)
 from rasa_core.domain import Domain
 from rasa_core.events import Restarted, SlotSet, UserUtteranceReverted
 from rasa_core.trackers import DialogueStateTracker
@@ -55,7 +55,7 @@ def test_domain_action_instantiation():
 
     instantiated_actions = domain.actions(None)
 
-    assert len(instantiated_actions) == 9
+    assert len(instantiated_actions) == 10
     assert instantiated_actions[0].name() == ACTION_LISTEN_NAME
     assert instantiated_actions[1].name() == ACTION_RESTART_NAME
     assert instantiated_actions[2].name() == ACTION_DEFAULT_FALLBACK_NAME
@@ -65,8 +65,9 @@ def test_domain_action_instantiation():
         ACTION_DEFAULT_ASK_AFFIRMATION_NAME)
     assert instantiated_actions[6].name() == (
         ACTION_DEFAULT_ASK_REPHRASE_NAME)
-    assert instantiated_actions[7].name() == "my_module.ActionTest"
-    assert instantiated_actions[8].name() == "utter_test"
+    assert instantiated_actions[7].name() == ACTION_BACK_NAME
+    assert instantiated_actions[8].name() == "my_module.ActionTest"
+    assert instantiated_actions[9].name() == "utter_test"
 
 
 def test_domain_fails_on_duplicated_actions():