+++ tests/test_channels.py
@@ -1,4 +1,5 @@
 import json
+import pytest
 
 from httpretty import httpretty
 
@@ -239,6 +240,9 @@ def test_rocketchat_channel():
 
 
 # USED FOR DOCS - don't rename without changing in the docs
+@pytest.mark.filterwarnings("ignore:"
+                            "unclosed file.*:"
+                            "ResourceWarning")
 def test_telegram_channel():
     # telegram channel will try to set a webhook, so we need to mock the api
 
@@ -282,6 +286,9 @@ def test_telegram_channel():
         httpretty.disable()
 
 
+@pytest.mark.filterwarnings("ignore:"
+                            "unclosed.*:"
+                            "ResourceWarning")
 def test_handling_of_telegram_user_id():
     # telegram channel will try to set a webhook, so we need to mock the api
 
@@ -568,6 +575,9 @@ def test_slackbot_init_two_parameter():
 
 
 # Use monkeypatch for sending attachments, images and plain text.
+@pytest.mark.filterwarnings("ignore:"
+                            "unclosed.*:"
+                            "ResourceWarning")
 def test_slackbot_send_attachment_only():
     from rasa_core.channels.slack import SlackBot
 
@@ -609,6 +619,9 @@ def test_slackbot_send_attachment_only():
                              'attachments': [attachment]}
 
 
+@pytest.mark.filterwarnings("ignore:"
+                            "unclosed.*:"
+                            "ResourceWarning")
 def test_slackbot_send_attachment_withtext():
     from rasa_core.channels.slack import SlackBot
 
@@ -653,6 +666,9 @@ def test_slackbot_send_attachment_withtext():
                              'attachments': [attachment]}
 
 
+@pytest.mark.filterwarnings("ignore:"
+                            "unclosed.*:"
+                            "ResourceWarning")
 def test_slackbot_send_image_url():
     from rasa_core.channels.slack import SlackBot
 
@@ -678,6 +694,9 @@ def test_slackbot_send_image_url():
            in r.parsed_body['attachments'][0]
 
 
+@pytest.mark.filterwarnings("ignore:"
+                            "unclosed.*:"
+                            "ResourceWarning")
 def test_slackbot_send_text():
     from rasa_core.channels.slack import SlackBot
 
@@ -698,6 +717,9 @@ def test_slackbot_send_text():
                              'text': ['my message']}
 
 
+@pytest.mark.filterwarnings("ignore:"
+                            "unclosed.*:"
+                            "ResourceWarning")
 def test_channel_inheritance():
     from rasa_core.channels import RestInput
     from rasa_core.channels import RasaChatInput