+++ tests/test_policies.py
@@ -20,6 +20,7 @@ from rasa_core.policies.embedding_policy import EmbeddingPolicy
 from rasa_core.policies.fallback import FallbackPolicy
 from rasa_core.policies.form_policy import FormPolicy
 from rasa_core.policies.keras_policy import KerasPolicy
+from rasa_core.policies.mapping_policy import MappingPolicy
 from rasa_core.policies.memoization import (
     MemoizationPolicy, AugmentedMemoizationPolicy)
 from rasa_core.policies.sklearn_policy import SklearnPolicy
@@ -121,6 +122,9 @@ class PolicyTestCollection(object):
         assert max(probabilities) <= 1.0
         assert min(probabilities) >= 0.0
 
+    @pytest.mark.filterwarnings("ignore:"
+                                ".*without a trained model present.*:"
+                                "UserWarning")
     def test_persist_and_load_empty_policy(self, tmpdir):
         empty_policy = self.create_policy(None, None)
         empty_policy.persist(tmpdir.strpath)
@@ -184,6 +188,14 @@ class TestFallbackPolicy(PolicyTestCollection):
 
         assert trained_policy.should_nlu_fallback(
             nlu_confidence, last_action_name) is should_nlu_fallback
+
+
+class TestMappingPolicy(PolicyTestCollection):
+
+    @pytest.fixture(scope="module")
+    def create_policy(self, featurizer, priority):
+        p = MappingPolicy()
+        return p
 
 
 class TestMemoizationPolicy(PolicyTestCollection):